Migrated from webpack to vite
