+++ package.json
@@ -17,7 +17,6 @@
     "react-dom": "^18.2.0",
     "react-feather": "^2.0.10",
     "react-router-dom": "^6.6.1",
-    "react-scripts": "5.0.1",
     "react-slick": "^0.29.0",
     "slick-carousel": "^1.8.1",
     "sort-by": "^1.2.0",
@@ -26,10 +25,9 @@
     "web-vitals": "^2.1.4"
   },
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject",
+    "start": "vite",
+    "build": "vite build",
+    "preview":"vite preview",
     "lint": "eslint ."
   },
   "eslintConfig": {
@@ -51,7 +49,8 @@
     ]
   },
   "devDependencies": {
+    "@vitejs/plugin-react": "^3.1.0",
     "json-loader": "^0.5.7",
-    "webpack-cli": "^5.0.1"
+    "vite": "^4.1.4"
   }
 }
