+++ src/website/OurDogs/OurDogs.css
@@ -4,7 +4,7 @@
   justify-content: center;
   align-items: center;
   margin: 20px 0px;
-  background-image: url("../../../public/images/dog-paw.jpg");
+  background-image: url("../../../images/dog-paw.jpg");
   height: 900px;
   background-repeat: repeat;
   background-size: 900px;
